Local LLM ready for initial review. Implemented the local llm feature, and migrated to an OOP style.
This is too much of a change but that is because I was developing this as separate thing. Essentially, it implements a local version of botex and at the moment does not implement the remote version. The LLMOTreeBotsManager class orchestrates the whole work flow by using the other classes. config.py deals with validating configurations, db_manager manages the database things, the message_handler deals with the llm side of things at the moment only the local but can extended to deal with remote as well, the web interaction and otree session managers also do their described action.