Implemented local llm, this time into the existing code :)
Implemented a local llm workflow for botex. This leaves the main workflow completely unaltered. There are improvements that I had to do on the other version to get the best out of the local llms, this changes will be pushed next, but I want to have a commit where the local llm is introduced with the limitations it has. This is mainly that it responds with index when presented with a select kind of question.