Initial test for local llm.
A first draft test for the local llm workflow. At the moment there is a lot to improve about it because the path to llama cpp, local model and number of layers to offload to gpu are hard coded into the test. But this is just a start and hopefully an improvement will be forthcoming.