Increase default context size, added more stopping seq, minor refactor.
I increased the default context size because the models that we would recommend have higher context size and I think llama cpp will deal with this if it is set too high, but too low would actually limit the model. Based on testing experience, I have added two more stopping sequences that indicate that the llm has gone of the rails. Then I droped the requirement for tokenizer.json, this was no longer needed as I do not need that for chat templating. Finally, I set has_system role to false if using default_chat_template asI am using the one from Mistral 7b_v3 as default and it does not have a system role.