Debug a 0 answer being marked as missing, debug the handling of correction conversation
A small edit to make sure that a 0 answer is not wrongly marked as missing answer. A bigger edit to make sure that the back and forth to get the right answer from the llm does not make it the conversation that is being recorded. At the moment this is not fully tested with gpt(gpt usually does not make mistake). Additionally, this part of the conversation is actually not recorded at all for now. This needs to be implemented down the line.