Remove debug text from page scrape and trim the text that is passed to llm.
First bit removed debug text from page scrape, this is when otree is started in debug mode. The debug information of otree clutters the message being passed to botex. Second bit trims the text that is passed to the llm. There is at times a lot of space at the end that gets scraped. Trimming that before passing it helps.