Taking off parallel llm analysis limitations
I had restrictions set in place to make sure that multiple instances of the llm model would not be started as that would have quickly overwhelmed the system. With the server multiple slots can be started to handle parallel requests. This also comes with potential to crush but this happens early and clearly so a user can fiddle around to figure out how many slots they can start safely depending on their machine.