Added the possibility for a user to pass prompts.
Now the user can provide user_prompts dictionary. This will be checked to make sure that the user is not providing a key that we are not expecting and if they are then, the program exits. Otherwise, the user prompt replaces any of the default prompts that we are using directly.