Added a separate history for botex_db, that is different from the conv_hist that is passed to the llm. Updated tests to reflect this.
Here we add a separate history for the botex_db, this is meant to ensure that all interaction with the llm make their way to the botex_db. On the other hand conv_hist which is passed to the llm remains functional, in that we try to avoid bloating the history by not passing all the error correction back and forth between the llm and botex. Additionally, the tests for the conversation being completed have to be updated so that they check the last accepted answer in botex_db, in case there has been some error handling.