Introduce LocalLLMConfig pydantic settings and pydantic ChatCompleitonResponse
First the LocalLLMConfig class is introduces. It does parameter validation, sets missing arguments from environment and if that is not available sets from default. It comes with additional dependency pydantic-settings but in my opinion it is worth it. Second, more refactoring and standardizing. A move away from the hacky MirrorLiteLLM class to a more standard ChatCompletionResponse class based on pydantic basemodel.