Introduce schemas, rename questions to answers (breaking)
This breaking change introduces schemas, that are sent along the prompt to the llm and almost guarantee the response format. They also allow validation as they are pydantic objects. It circumvents the need for rigors checking and retrying. The breaking part is that only models that support structured output will be supported, these are gpt-4o-mini-2024-07-18 and later or gpt-4o-2024-08-06 and later. For the local llm, "recent" builds of llama.cpp and more recent llms work very well.

Another change that was done was to rename the object that was received from the llm from question to answer, because what we are receiving from the llm is answers and not questions. The old name makes the creation of the answer response model schema difficult to grasp (at least for me). Finally, to further aid in the creation of the answer model, the data structure of answers was changed from list ['id', 'reason', 'answer'] to a dict {'id': {'reason', 'answer'}, this made it easier to ensure multiple answers to the same question were not being returned.