Change questions to a dict, debug local llm, extend schemas
Here I update the data structure of questions to a dictionary instead of a list following the change in the data structure of the answer expected from the llm. I also remove some bugs in the local llm class that were surfacing when the set_params_from_running_api tries to set some specific parameters (like temperature ...) instead of passing from config. Finally, I extend the schema by adding field descriptions. Arguably this is not fully utilized yet because I am not sure if the field descriptions are used by either OpenAI or llama.cpp but I have them for the sake of completion.